// src/models/models.tsx
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]?: Maybe<T[SubKey]>;
};
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]: Maybe<T[SubKey]>;
};
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  Time: any;
  Upload: any;
};
// ...
export type ChangePasswordInput = {
  previousPassword: Scalars["String"];
  password: Scalars["String"];
  confirmPassword: Scalars["String"];
};
// ...
export type FileUpload = {
  file: Scalars["Upload"];
  name: Scalars["String"];
};
// ...
export type MutationUpdateUserArgs = {
  input: UserUpdateInput;
};

export type MutationChangePasswordArgs = {
  input: ChangePasswordInput;
};
// ...
export type QueryUserArgs = {
  id: Scalars["ID"];
};
// ...
export type UserUpdateInput = {
  id: Scalars["ID"];
  firstName: Scalars["String"];
  lastName: Scalars["String"];
  email: Scalars["String"];
  userTypeIds: Array<InputMaybe<Scalars["ID"]>>;
  active?: InputMaybe<Scalars["Boolean"]>;
  signature?: InputMaybe<FileUpload>;
  profilePic?: InputMaybe<FileUpload>;
};